% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{extract_genes_in_path}
\alias{extract_genes_in_path}
\title{Return observed genes in a given pathway}
\usage{
extract_genes_in_path(correlation_data, gsea_path, gsea_path_db)
}
\description{
Return observed genes in a given pathway
}
