% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{gen_correlation_matrix}
\alias{gen_correlation_matrix}
\title{Generates list of different correlation matrices}
\usage{
gen_correlation_matrix(data, methods = c("pearson", "spearman"))
}
\arguments{
\item{data}{output of the prep_msstats_data function.}

\item{methods}{list of correlation methods to use (in cor function)}
}
\description{
Takes the output of the prep_msstats_data function as input
}
