% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{prep_msstats_data}
\alias{prep_msstats_data}
\title{Prepares MS label-free data for GSEA correlation analysis}
\usage{
prep_msstats_data(data, gene_map = NULL, parse_gene = FALSE)
}
\arguments{
\item{data}{The ProteinLevelData object returned from the MSstats dataProcess
function.}

\item{gene_map}{table which maps uniprot protein names to gene names, if
required.}

\item{parse_genes}{if gene name has a "_" (e.g. SPTBN2_HUMAN), setting this
to TRUE will remove the string after the underscore.}
}
\description{
Takes the ProteinLevel output of the MSstats dataProcess function as input.
}
