% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{test_GSEA_pathways}
\alias{test_GSEA_pathways}
\title{Checks correlations between genes in GSEA pathways}
\usage{
test_GSEA_pathways(correlation_data, gsea_path, threshold = 0.5)
}
\arguments{
\item{correlation_data}{output of gen_correlation_matrix function}

\item{gsea_path}{file path to GSEA JSON file}

\item{threshold}{correlation threshold to count as significant correlation}
}
\description{
takes as input correlation list and gsea file path
}
